refactor(api): move activity step tests to workflow-level integration tests
Test behavior through the workflow instead of testing addActivitiesStep
directly, so tests survive implementation changes.
